Fix the defects the new warning set found
Several of the Tier A warnings pointed at real bugs rather than style:

- deviceFloppyPIO.c: the head position is unsigned, so stepping out past
  track 0 wrapped round and was then clamped to track 76 instead of 0.
  The step is now computed signed and clamped to 0..76.
- deviceSMD.c: the write path copied Device_DMARead()'s int32_t result into
  a uint32_t, so its "readData < 0" DMA-error test could never be true.
  The compare path now also treats a failed DMA read explicitly.
- menu.c: print_wrapped_text() copied a word wider than the line into a
  line-sized stack array with strcpy (buffer overflow), and the truncating
  helpers built a VLA of available_width + 4 bytes, which is negative for
  a narrow window. Printing now uses %.*s and a fixed 1024-byte line.
  strdup() in print_wrapped_text() is checked.
- cpu_disasm.c: the two ND-110 sub-opcode switches fall through into the
  next case on purpose; that is now marked. The printable-character test
  used '&' where '&&' was meant, on a signed char.
- cpu.c calcPK(): the level counter was unsigned, making "lvl >= 0" always
  true. It could not loop forever only because the caller guarantees a set
  bit; it is now an int.

The rest are mechanical: const on the disassembler string tables and
GetPageTableEntryDebugInfo(), explicit casts where a negative value is
already excluded, parentheses in panel.c, #ifdef for three never-defined
dead-code macros, a "/*" inside a comment, and larger path buffers in
test_printjob.c. All 16 ctest tests pass.

+++ src/cpu/cpu.c
@@ -407,7 +407,7 @@ void calcPK(void)
 {
 	// Recalculate PK based on PID and PIE
 	int s;
-	ushort lvl;
+	int lvl;
 	ushort i;
 	gPK = 0;
 	i = gPIE & gPID;
